renamed file to teddy.js and passes linting
